Citalac functions and fixing mistakes

+++ WEB/src/main/java/com/INprojekat/WEB/service/PolicaService.java
@@ -44,6 +44,28 @@ public class PolicaService {
         return null;
     }
 
+
+    public Polica findOneById(Long id){
+        Optional<Polica> foundPolica = policaRepository.findById(id);
+        if (foundPolica.isPresent()) {
+            return foundPolica.get();
+        }
+        return null;
+    }
+    public List<Polica> findAll(){
+        return policaRepository.findAll();
+    }
+
+    public Set<PolicaDto> getPolice(Long id) {
+        Korisnik korisnik = korisnikService.findOne(id);
+        Set<Polica> police = korisnik.getPolice();
+        Set<PolicaDto> dtos = new HashSet<>();
+        for(Polica polica: police){
+            PolicaDto dto = new PolicaDto(polica);
+        }
+        return dtos;
+    }
+
     public void create(PolicaAddDto policaAddDto, Long id) {
         Polica polica;
         polica = new Polica();
@@ -58,53 +80,61 @@ public class PolicaService {
     }
 
     public void main3(){
-        Polica WantToRead = new Polica();
-        WantToRead.setNaziv("Want To Read");
-        WantToRead.setPrimarna(true);
-        save(WantToRead);
+            Polica WantToRead = new Polica();
+            WantToRead.setNaziv("Want To Read");
+            WantToRead.setPrimarna(true);
+            save(WantToRead);
 
-        Polica CurrentlyReading = new Polica();
-        CurrentlyReading.setNaziv("CurrentlyReading");
-        CurrentlyReading.setPrimarna(true);
-        save(CurrentlyReading);
+            Polica CurrentlyReading = new Polica();
+            CurrentlyReading.setNaziv("CurrentlyReading");
+            CurrentlyReading.setPrimarna(true);
+            save(CurrentlyReading);
 
-        Polica Read = new Polica();
-        Read.setNaziv("Read");
-        Read.setPrimarna(true);
-        save(Read);
+            Polica Read = new Polica();
+            Read.setNaziv("Read");
+            Read.setPrimarna(true);
+            save(Read);
     }
 
-    public List<Polica> findAll(){
-        return policaRepository.findAll();
-    }
 
-    public Boolean existsPolica(String naziv) { return policaRepository.existsByNaziv(naziv); }
-
-    public Polica save(Polica polica) { return policaRepository.save(polica);}
-
-    public void deletePolica(Long id) throws ChangeSetPersister.NotFoundException {
-        //potrebna provera da li je primarna =, ako je primarna vrati EXCEPTION da se ne moze obrisati, a ako nije primarna, obirsati je
-        Polica polica = policaRepository.findById(id)
-                .orElseThrow(() -> new ChangeSetPersister.NotFoundException());
-        policaRepository.delete(polica);
-    }
-    public void addKnjigaOnPolica(Long policaId, Long knjigaId) throws ChangeSetPersister.NotFoundException {
+    public Long addKnjigaOnPolica(Long policaId, Long knjigaId) throws ChangeSetPersister.NotFoundException {
         Polica polica = policaRepository.findById(policaId).orElse(null);
         Knjiga knjiga = knjigaRepository.findById(knjigaId).orElse(null);
         StavkaPolice stavkaPolice = new StavkaPolice();
+
         stavkaPolice.setKnjiga(knjiga);
         stavkaPoliceRepository.save(stavkaPolice);
         polica.getStavkePolica().add(stavkaPolice);
         save(polica);
 
+        return stavkaPolice.getId();
     }
-    public Set<PolicaDto> getPolice(Long id) {
-        Korisnik korisnik = korisnikService.findOne(id);
-        Set<Polica> police = korisnik.getPolice();
-        Set<PolicaDto> dtos = new HashSet<>();
-        for(Polica polica: police){
-            PolicaDto dto = new PolicaDto(polica);
+    public boolean deletePolica(Long citalac_autorId,Long policaId) throws ChangeSetPersister.NotFoundException {
+        Polica polica = policaRepository.findById(policaId)
+                .orElseThrow(() -> new ChangeSetPersister.NotFoundException());
+        if(polica.isPrimarna() == false) {
+            Korisnik korisnik = korisnikService.findOne(citalac_autorId);
+            Set<Polica> police = korisnik.getPolice();
+            police.remove(polica);
+            policaRepository.delete(polica);
+            korisnik.setPolice(police);
+            korisnikService.save(korisnik);
+            return true;
         }
-        return dtos;
+        return false;
     }
+    public Boolean existsPolica(String naziv) { return policaRepository.existsByNaziv(naziv); }
+    public Polica save(Polica polica) { return policaRepository.save(polica);}
+
+/*
+        1. Trazenje jedne police
+        2. Izlistavanje svih polica
+        3. Izlistavanje svih polica korisnika
+        4. Kreiranje nove police
+        5. Kreiranje 3 main police
+        6. Dodavanje knjige na policu
+        7. Brisanje police
+        8. Trazenje polica istog imena
+        9. Cuvanje police
+ */
 }
